Fixed error with functions decl vs defn having only different return types (int64-t => ssize-t).

+++ include/dakota.hh
@@ -265,8 +265,8 @@ constexpr FUNC dk_hash_switch(size_t  val) -> size_t  { return val; }
 
 [[debug_export]] FUNC dkt_get_klass_chain(object_t klass, char8_t* buf, int64_t buf_len) -> char8_t*;
 
-[[debug_export]] FUNC dkt_dump_methods(object_t)        -> int64_t;
-[[debug_export]] FUNC dkt_dump_methods(klass::slots_t*) -> int64_t;
+[[debug_export]] FUNC dkt_dump_methods(object_t)        -> ssize_t;
+[[debug_export]] FUNC dkt_dump_methods(klass::slots_t*) -> ssize_t;
 
 [[debug_export]] FUNC dkt_unbox_check(object_t object, object_t kls) -> void;
